64-bit OSDText
We'll hopefully never ever need OSD text widgets that hold more than 4GB of
text data. Though the pattern in the code is (on 64-bit platforms):
 a) start from 64-bit values (offsets and sizes in string objects)
 b) do some calculations on those values (e.g. find a good position for where
    to split a string to display it over multiple lines)
 c) use the result of this calculation as offset/length in a string (those
    string methods expect 64-bit values)

Before this patch we would between step a and b truncate the 64 bit values to
32 bit. And between step b and c zero-extend the values back to 64 bit. The
code actually becomes simpler if we don't do these conversions. (Although on
x86 32<->64 bit truncation/extension operations are very cheap).

Note that (like all earlier patches in this series) on 32-bit systems nothing
changes. So on 32-bit systems the intermediate calculations are still performed
using 32 bit. (In other words, only if 64-bit operations are efficient, we use
those).

git-svn-id: file:///openmsx-svn/code/openmsx/trunk@13091 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/console/OSDText.hh
@@ -33,7 +33,7 @@ private:
 		OutputRectangle& output);
 
 	template<typename FindSplitPointFunc, typename CantSplitFunc>
-	unsigned split(const std::string& line, unsigned maxWidth,
+	size_t split(const std::string& line, unsigned maxWidth,
 		FindSplitPointFunc findSplitPoint, CantSplitFunc cantSplit,
 		bool removeTrailingSpaces) const;
 	unsigned splitAtChar(const std::string& line, unsigned maxWidth) const;
